Refactored style arrow
